+++ packages/kbn-optimizer/limits.yml
@@ -178,7 +178,7 @@ pageLoadAssetSize:
   unifiedHistogram: 19928
   unifiedSearch: 71059
   upgradeAssistant: 81241
-  uptime: 37842
+  uptime: 60000
   urlDrilldown: 30063
   urlForwarding: 32579
   usageCollection: 39762
