+++ x-pack/solutions/observability/plugins/uptime/public/locators/overview.ts
@@ -11,21 +11,8 @@ import {
   type UptimeOverviewLocatorParams,
 } from '@kbn/deeplinks-observability';
 import type { LocatorDefinition } from '@kbn/share-plugin/common';
-import { OVERVIEW_ROUTE } from '../../common/constants';
 
 export type { UptimeOverviewLocatorInfraParams, UptimeOverviewLocatorParams };
-
-const formatSearchKey = (key: string, value: string) => `${key}: "${value}"`;
-
-function isUptimeOverviewLocatorParams(
-  args: UptimeOverviewLocatorInfraParams | UptimeOverviewLocatorParams
-): args is UptimeOverviewLocatorParams {
-  return (
-    (args as UptimeOverviewLocatorParams).search !== undefined ||
-    (args as UptimeOverviewLocatorParams).dateRangeEnd !== undefined ||
-    (args as UptimeOverviewLocatorParams).dateRangeStart !== undefined
-  );
-}
 
 export class UptimeOverviewLocatorDefinition
   implements LocatorDefinition<UptimeOverviewLocatorInfraParams | UptimeOverviewLocatorParams>
@@ -35,36 +22,7 @@ export class UptimeOverviewLocatorDefinition
   public readonly getLocation = async (
     params: UptimeOverviewLocatorInfraParams | UptimeOverviewLocatorParams
   ) => {
-    let qs = '';
-
-    if (isUptimeOverviewLocatorParams(params)) {
-      qs = Object.entries(params)
-        .map(([key, value]) => {
-          if (value) {
-            return `${key}=${value}`;
-          }
-        })
-        .join('&');
-    } else {
-      const searchParams: string[] = [];
-      if (params.host) searchParams.push(formatSearchKey('host.name', params.host));
-      if (params.container) searchParams.push(formatSearchKey('container.id', params.container));
-      if (params.pod) searchParams.push(formatSearchKey('kubernetes.pod.uid', params.pod));
-      if (params.ip) {
-        searchParams.push(formatSearchKey(`host.ip`, params.ip));
-        searchParams.push(formatSearchKey(`monitor.ip`, params.ip));
-      }
-      if (searchParams.length > 0) {
-        qs = `search=${searchParams.join(' OR ')}`;
-      }
-    }
-
-    const path = `${OVERVIEW_ROUTE}${qs ? `?${qs}` : ''}`;
-
-    return {
-      app: 'uptime',
-      path,
-      state: {},
-    };
+    const { getLocation } = await import('./get_location');
+    return getLocation(params);
   };
 }
